Rename tensor.py to wrapper_tensor.py. Rename quantize.py to transform.py. Update imports in other files.
