+++ test/prototype/moe_qat/test_distributed.py
@@ -13,7 +13,7 @@ from torch.nn import functional as F
 
 from torchao.float8.float8_utils import compute_error
 from torchao.prototype.moe_qat import MoEQATConfig
-from torchao.prototype.moe_qat.tensor import FakeQuantizedWeightWrapperBaseTensor, Float8FakeQuantizedWeightWrapperTensor
+from torchao.prototype.moe_qat.wrapper_tensor import FakeQuantizedWeightWrapperBaseTensor, Float8FakeQuantizedWeightWrapperTensor
 from torchao.quantization.qat.fake_quantize_config import Float8FakeQuantizeConfig
 from torchao.quantization.granularity import PerRow
 from torchao.quantization.quant_api import quantize_
